refactor: split renpho API, formatters, and MCP tool registration
Census-ranked targets (renpho-api B3 mixed concerns, formatting grab-bag)
split into focused modules with façade re-exports preserving import paths.
MCP tool handlers moved to register-tools.ts; behavior unchanged.
